Don't include HTTP body in GET request
A body is technically allowed in a HTTP GET request, but it must not
have any meaning.  We have been accidentally including a two byte body
of "{}" for all GET requests generated by the ruby gem.  This body is
included when signing the request, so omitting the body (for example
when passing the API request through a proxy or HTTP redirect) will
cause server-side signature verification code to fail to authenticate
the request.  To fix this, don't force a default of {}.to_json when no
payload is passed to the request.