Change spec redirect code to the real one
We started using "307 Temporary Redirect" since it is defined as not
allowing the client to change the HTTP method, which is the behavior we
want, and some clients will only exhibit that behavior with 307/308.
The ruby client works either way, but make sure it's tested with the
more realistic 307 status code.